RequestParser as abstract base class with JSON & XML implementations
